Issue 609 - Enhance InferBound

+++ applications/newton/llvm-ir/c-files/math_config.h
@@ -117,7 +117,7 @@ converttoint (double_t x)
 #endif
 
 static inline uint32_t
-asuint (float f)
+asuint (double f)
 {
 	union
 	{
